# functions/docs/swagger.yaml
openapi: "3.0.0"
info:
  title: "전기안전 - Firebase Functions API"
  version: "1.0.0"
  description: |
    ## 🚀 Endpoint
    배포가 완료되면 아래 주소에서 API를 확인할 수 있습니다.

    - `https://us-central1-electrical-safety-4c9bd.cloudfunctions.net/api`
      <br>  <br>
    - `https://api-2hj4xnf47q-uc.a.run.app`
paths:
...
  /sensor/{sensorId}:
    get:
      tags:
        - Sensors
      summary: 특정 센서의 모든 데이터 조회
      parameters:
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
          description: 센서의 고유 ID
      responses:
        "200":
          description: 센서 데이터 반환
          content:
            application/json:
              schema:
                type: object
        "404":
          description: 데이터가 없을 경우
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "500":
          description: 서버 오류
  /sensor/{sensorId}/latest:
    get:
      tags:
        - Sensors
      summary: 특정 센서의 최신 데이터 조회
      parameters:
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 최신 데이터 반환
          content:
            application/json:
              schema:
                type: object
        "404":
          description: 데이터가 없을 경우
        "500":
          description: 서버 오류
  /sensor/{sensorId}/daily-average/{date}:
    get:
      tags:
        - Sensors
      summary: 특정 센서의 일간 평균 데이터 조회
      parameters:
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
          description: YYYY-MM-DD 형식의 날짜
      responses:
        "200":
          description: 전체 평균 및 시간별 평균 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  overallAvg:
                    type: number
                  hourlyAvg:
                    type: object
                    additionalProperties:
                      type: number
        "404":
          description: 데이터가 없을 경우
        "500":
          description: 서버 오류
  /sensor/{sensorId}/monthly-average/{year}/{month}:
    get:
      tags:
        - Sensors
      summary: 특정 센서의 월간 평균 데이터 조회
      parameters:
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
        - in: path
          name: year
          required: true
          schema:
            type: integer
        - in: path
          name: month
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        "200":
          description: 전체 평균 및 주차별 평균 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  overallAvg:
                    type: number
                  weeklyAvg:
                    type: object
                    additionalProperties:
                      type: number
        "404":
          description: 데이터가 없을 경우
        "500":
          description: 서버 오류
  /sensor/{sensorId}/yearly-average/{year}:
    get:
      tags:
        - Sensors
      summary: 특정 센서의 연간 평균 데이터 조회
      parameters:
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
        - in: path
          name: year
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 전체 평균 및 월별 평균 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  overallAvg:
                    type: number
                  monthlyAvg:
                    type: object
                    additionalProperties:
                      type: number
        "404":
          description: 데이터가 없을 경우
        "500":
          description: 서버 오류

    components:
      schemas:
        TemperatureData:
          type: object
          properties:
            sensorId:
              type: string
            temperature:
              type: number
            timestamp:
              type: string
              format: date-time
  /sensor/users/{userId}/sensors:
    post:
      tags:
        - Sensors(유저)
      summary: 센서 키 등록
      description: 사용자의 센서 키를 등록합니다.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 사용자 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sensorId
              properties:
                sensorId:
                  type: string
                  example: sensor123
      responses:
        "200":
          description: 센서 키가 등록되었습니다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  sensorId:
                    type: string
        "400":
          description: 요청 바디 오류
        "500":
          description: 서버 오류

    get:
      tags:
        - Sensors(유저)
      summary: 센서 키 목록 조회
      description: 사용자가 등록한 센서 키 목록을 조회합니다.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 사용자 ID
      responses:
        "200":
          description: 센서 키 목록 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  sensors:
                    type: array
                    items:
                      type: string
        "404":
          description: 센서 키가 없습니다.
        "500":
          description: 서버 오류
  /sensor/users/{userId}/sensors/{sensorId}:
    delete:
      tags:
        - Sensors(유저)
      summary: 센서 키 삭제
      description: 사용자의 특정 센서 키를 삭제합니다.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: 사용자 ID
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
          description: 삭제할 센서 ID
      responses:
        "200":
          description: 센서 키 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  sensorId:
                    type: string
        "404":
          description: 센서 키를 찾을 수 없음
        "500":
          description: 서버 오류
...